Refactor Redirects Out of Node Lifecycle
And put redirect events in the web server.
