+++ src/server/oauth_installed_flow_delegate.rs
@@ -1,5 +1,7 @@
 use crate::domain::mpsc_handle::one_shot;
 use crate::domain::node::Manager;
+use crate::server::Event::Redirect;
+use crate::server::WebEventChannelHandle;
 use log::debug;
 use std::future::Future;
 use std::pin::Pin;
@@ -11,28 +13,32 @@ use yup_oauth2::authenticator_delegate::InstalledFlowDelegate;
 #[derive(Clone, Debug)]
 pub struct OauthFlowDelegateFactory {
     node_manager: Manager,
+    web_channel_handle: WebEventChannelHandle,
 }
 
 impl OauthFlowDelegateFactory {
-    pub fn new(manager: &Manager) -> Self {
+    pub fn new(manager: &Manager, web_channel_handle: &WebEventChannelHandle) -> Self {
         Self {
             node_manager: manager.clone(),
+            web_channel_handle: web_channel_handle.clone(),
         }
     }
 
     pub fn get_installed_flow_delegate(&self) -> OauthInstalledFlowDelegate {
-        OauthInstalledFlowDelegate::new(&self.node_manager)
+        OauthInstalledFlowDelegate::new(&self.node_manager, &self.web_channel_handle)
     }
 }
 
 pub struct OauthInstalledFlowDelegate {
     node_manager: Manager,
+    web_channel_handle: WebEventChannelHandle,
 }
 
 impl OauthInstalledFlowDelegate {
-    pub fn new(manager: &Manager) -> Self {
+    pub fn new(manager: &Manager, web_channel_handle: &WebEventChannelHandle) -> Self {
         Self {
             node_manager: manager.clone(),
+            web_channel_handle: web_channel_handle.clone(),
         }
     }
 }
@@ -43,12 +49,18 @@ impl InstalledFlowDelegate for OauthInstalledFlowDelegate {
         url: &'a str,
         need_code: bool,
     ) -> Pin<Box<dyn Future<Output = Result<String, String>> + Send + 'a>> {
-        Box::pin(present_user_url(&self.node_manager, url, need_code))
+        Box::pin(present_user_url(
+            &self.node_manager,
+            &self.web_channel_handle,
+            url,
+            need_code,
+        ))
     }
 }
 
 async fn present_user_url(
     node_manager: &Manager,
+    web_channel_handle: &WebEventChannelHandle,
     url: &str,
     need_code: bool,
 ) -> Result<String, String> {
@@ -71,9 +83,9 @@ async fn present_user_url(
         .await
         .expect("Could not acquire semaphore");
 
-    node_manager
+    web_channel_handle
         .clone()
-        .send_oauth2_redirect_url(url, need_code, sender)
+        .send(Redirect(url.to_string(), sender))
         .map_err(|e| format!("Failed to send oauth2 redirect url: {}", e))?;
 
     let result = task.await;
