+++ src/server/websocket/websocket.rs
@@ -1,5 +1,6 @@
 use crate::core::node_handles::NodeHandles;
 use crate::domain::node::{InitReplier, Lifecycle};
+use crate::server::Event::Redirect;
 use log::{debug, error, info};
 use std::future;
 use std::pin::Pin;
@@ -7,7 +8,6 @@ use tokio::sync::broadcast::error::RecvError;
 use tokio::task;
 use warp::ws::{Message, WebSocket};
 use warp::{path, ws, Filter, Rejection, Reply, Sink};
-use Lifecycle::{Redirect, Stop};
 
 type Result<T> = std::result::Result<T, Rejection>;
 
@@ -36,23 +36,30 @@ async fn use_websocket(mut web_socket: WebSocket, handles: NodeHandles) {
     let handles = handles.clone();
 
     let lifecycle_reader = handles.lifecycle_manager().readonly();
+    let web_event_reader = handles.web_channel_handle().clone();
 
-    task::spawn(async move {
+    let task = task::spawn(async move {
         loop {
-            match lifecycle_reader.get_receiver().recv().await {
+            match web_event_reader.get_receiver().recv().await {
                 Ok(event) => match event {
                     Redirect(url, sender) => {
+                        // TODO - retry on timeout
                         let message = Message::text(format!(
                             "{{\"type\":\"redirect_event\",\"url\":\"{:?}\"}}",
                             url
                         ));
+
+                        debug!("Polling websocket ready.");
                         let mut pinned_socket = Pin::new(&mut web_socket);
                         future::poll_fn(|cx| pinned_socket.as_mut().poll_ready(cx))
                             .await
                             .unwrap();
+
+                        debug!("Sending redirect event: {:?}", message);
                         let mut pinned_socket = Pin::new(&mut web_socket);
                         pinned_socket.as_mut().start_send(message).unwrap();
 
+                        debug!("Polling websocket flush.");
                         future::poll_fn(|cx| pinned_socket.as_mut().poll_flush(cx))
                             .await
                             .unwrap();
@@ -64,12 +71,9 @@ async fn use_websocket(mut web_socket: WebSocket, handles: NodeHandles) {
                             Err(_) => {
                                 error!("Redirect websocket event failed to send.");
                             }
-                        }
-                    }
-                    Stop => {
+                        };
                         break;
                     }
-                    _ => {}
                 },
                 Err(e) => match e {
                     RecvError::Closed => {
@@ -82,4 +86,6 @@ async fn use_websocket(mut web_socket: WebSocket, handles: NodeHandles) {
             }
         }
     });
+
+    lifecycle_reader.abort_on_stop(&task);
 }
